backend/requirements.txt
# Requirements placeholder 
fastapi
uvicorn
pytesseract
Pillow
pdf2image
chromadb
sentence-transformers 